booking - move payment to a separate store, remove unused files, prevent error in choose dates when no roomtype
