--- src/components/Reservation/Reservation.vue
@@ -1,17 +1,0 @@
-<template>
-  <router-view></router-view>
-</template>
-
-<script>
-const ReviewRules = () => import ('@/components/Reservation/ReviewRules.vue')
-
-export default {
-  components:{
-    ReviewRules
-  }
-}
-</script>
-
-<style scoped>
-
-</style>
